requirements.txt
numpy==1.19.5
kafka-python==2.0.2